infered userId

+++ app/src/main/java/com/mylifemobile/api/model/ExpensesModel.kt
@@ -6,5 +6,5 @@ import java.util.*
 data class ExpensesModel (@SerializedName("id") var id:Int = 0,
                           @SerializedName("amount") var amount:Number,
                           @SerializedName("date") var date: Date,
-                          @SerializedName("user_id") var userId: Long,
-                          @SerializedName("category_id") var categoryId: Long)
+                          @SerializedName("user_id") var userId: Int,
+                          @SerializedName("category_id") var categoryId: Int)
